allow children to be character vector of node names
variable 'children' is documented to accept only an integer argument for the number of parallel processes. but underlying psock process can use a character vector of node names instead of an integer, to parallellize across multiple nodes. parallel_functions.R was modified so that if such a vector is passed, the length of the vector is used in cases where the number of processes is needed